+++ modules/core/src/main/java/com/developmentontheedge/be5/modules/core/operations/users/Login.java
@@ -5,7 +5,7 @@ import com.developmentontheedge.be5.meta.UserAwareMeta;
 import com.developmentontheedge.be5.modules.core.services.LoginService;
 import com.developmentontheedge.be5.operation.OperationResult;
 import com.developmentontheedge.be5.server.FrontendActions;
-import com.developmentontheedge.be5.server.helpers.UserHelper;
+import com.developmentontheedge.be5.server.services.users.UserHelper;
 import com.developmentontheedge.be5.server.operations.support.GOperationSupport;
 import com.developmentontheedge.be5.server.services.UserInfoModelService;
 import com.developmentontheedge.beans.DPBuilder;
