+++ modules/core/src/main/java/com/developmentontheedge/be5/modules/core/operations/users/Logout.java
@@ -1,7 +1,7 @@
 package com.developmentontheedge.be5.modules.core.operations.users;
 
 import com.developmentontheedge.be5.server.FrontendActions;
-import com.developmentontheedge.be5.server.helpers.UserHelper;
+import com.developmentontheedge.be5.server.services.users.UserHelper;
 import com.developmentontheedge.be5.server.operations.support.GOperationSupport;
 import com.developmentontheedge.be5.server.services.UserInfoModelService;
