+++ modules/core/src/main/java/com/developmentontheedge/be5/modules/core/services/RoleHelperImpl.java
@@ -4,7 +4,7 @@ import com.developmentontheedge.be5.config.CoreUtils;
 import com.developmentontheedge.be5.database.DbService;
 import com.developmentontheedge.be5.metadata.DatabaseConstants;
 import com.developmentontheedge.be5.metadata.MetadataUtils;
-import com.developmentontheedge.be5.server.helpers.RoleHelper;
+import com.developmentontheedge.be5.server.services.users.RoleHelper;
 
 import javax.inject.Inject;
 import java.util.ArrayList;
