+++ server/src/main/java/com/developmentontheedge/be5/server/ServerModule.java
@@ -8,11 +8,11 @@ import com.developmentontheedge.be5.databasemodel.helpers.SqlHelper;
 import com.developmentontheedge.be5.operation.OperationModule;
 import com.developmentontheedge.be5.query.QueryModule;
 import com.developmentontheedge.be5.query.QuerySession;
-import com.developmentontheedge.be5.server.helpers.DpsHelper;
-import com.developmentontheedge.be5.server.helpers.ErrorModelHelper;
-import com.developmentontheedge.be5.server.helpers.FilterHelper;
-import com.developmentontheedge.be5.server.helpers.MenuHelper;
-import com.developmentontheedge.be5.server.helpers.UserHelper;
+import com.developmentontheedge.be5.server.services.DpsHelper;
+import com.developmentontheedge.be5.server.services.ErrorModelHelper;
+import com.developmentontheedge.be5.server.services.FilterHelper;
+import com.developmentontheedge.be5.server.services.MenuHelper;
+import com.developmentontheedge.be5.server.services.users.UserHelper;
 import com.developmentontheedge.be5.server.services.DocumentFormPlugin;
 import com.developmentontheedge.be5.server.services.DocumentGenerator;
 import com.developmentontheedge.be5.server.services.DocumentOperationsPlugin;
