+++ server/src/main/java/com/developmentontheedge/be5/server/operations/FilterOperation.java
@@ -1,6 +1,6 @@
 package com.developmentontheedge.be5.server.operations;
 
-import com.developmentontheedge.be5.server.helpers.FilterHelper;
+import com.developmentontheedge.be5.server.services.FilterHelper;
 import com.developmentontheedge.be5.server.operations.support.OperationSupport;
 import com.developmentontheedge.be5.util.FilterUtil;
 import com.developmentontheedge.beans.DynamicPropertySet;
