+++ server/src/main/java/com/developmentontheedge/be5/server/operations/support/OperationExtenderSupport.java
@@ -9,7 +9,7 @@ import com.developmentontheedge.be5.operation.services.OperationBuilder;
 import com.developmentontheedge.be5.operation.validation.Validator;
 import com.developmentontheedge.be5.operation.support.BaseOperationExtenderSupport;
 import com.developmentontheedge.be5.query.services.QueriesService;
-import com.developmentontheedge.be5.server.helpers.DpsHelper;
+import com.developmentontheedge.be5.server.services.DpsHelper;
 import com.developmentontheedge.be5.web.Request;
 import com.developmentontheedge.be5.web.Session;
 import com.developmentontheedge.be5.web.impl.FileUploadWrapper;
