+++ server/src/main/java/com/developmentontheedge/be5/server/services/FilterHelper.java
@@ -1,4 +1,4 @@
-package com.developmentontheedge.be5.server.helpers;
+package com.developmentontheedge.be5.server.services;
 
 import com.developmentontheedge.be5.databasemodel.util.DpsUtils;
 import com.developmentontheedge.be5.metadata.model.Query;
