+++ server/src/main/java/com/developmentontheedge/be5/server/services/users/RoleHelper.java
@@ -1,4 +1,4 @@
-package com.developmentontheedge.be5.server.helpers;
+package com.developmentontheedge.be5.server.services.users;
 
 import java.util.Collection;
 import java.util.List;
