+++ server/src/main/java/com/developmentontheedge/be5/server/servlet/TemplateFilter.java
@@ -2,7 +2,7 @@ package com.developmentontheedge.be5.server.servlet;
 
 import com.developmentontheedge.be5.meta.ProjectProvider;
 import com.developmentontheedge.be5.security.UserInfoHolder;
-import com.developmentontheedge.be5.server.helpers.UserHelper;
+import com.developmentontheedge.be5.server.services.users.UserHelper;
 import com.developmentontheedge.be5.server.services.HtmlMetaTags;
 import com.developmentontheedge.be5.server.servlet.support.FilterSupport;
 import com.developmentontheedge.be5.server.servlet.support.ServletUtils;
