+++ server/src/test/groovy/com/developmentontheedge/be5/server/helpers/DpsHelperCustomOpTest.groovy
@@ -1,6 +1,7 @@
 package com.developmentontheedge.be5.server.helpers
 
 import com.developmentontheedge.be5.groovy.GDynamicPropertySetSupport
+import com.developmentontheedge.be5.server.services.DpsHelper
 import com.developmentontheedge.be5.test.ServerBe5ProjectTest
 import org.junit.Before
 import org.junit.Test
