+++ server/src/test/java/com/developmentontheedge/be5/server/helpers/MenuHelperTest.java
@@ -3,6 +3,7 @@ package com.developmentontheedge.be5.server.helpers;
 import com.developmentontheedge.be5.metadata.RoleType;
 import com.developmentontheedge.be5.metadata.model.EntityType;
 import com.developmentontheedge.be5.server.model.Action;
+import com.developmentontheedge.be5.server.services.MenuHelper;
 import com.developmentontheedge.be5.test.ServerBe5ProjectTest;
 import org.junit.Test;
