+++ server/src/test/java/com/developmentontheedge/be5/test/ServerBe5ProjectDBTest.java
@@ -4,7 +4,7 @@ import com.developmentontheedge.be5.cache.Be5Caches;
 import com.developmentontheedge.be5.meta.ProjectProvider;
 import com.developmentontheedge.be5.server.ServerModule;
 import com.developmentontheedge.be5.server.services.rememberme.PersistentTokenRepository;
-import com.developmentontheedge.be5.server.helpers.RoleHelper;
+import com.developmentontheedge.be5.server.services.users.RoleHelper;
 import com.developmentontheedge.be5.server.services.InitUserService;
 import com.developmentontheedge.be5.server.services.OperationLogging;
 import com.developmentontheedge.be5.server.services.events.Be5EventTestLogger;
